LayoutTests/media/W3C/video/paused/paused_false_during_play-expected.txt

video.paused should be false during play event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS false is false

TEST COMPLETE
spec reference
